Update metrics to use productcatalogservice
Frontend CLIENT spans are no longer available, so modifying the labs to create SLO for the product catalog service, instead. This makes the SLO concept simpler to understand, monitoring just one simple service, which has two spans. In the environment, productcatalog is configured to generate errors, as per the feature toggle.